TIPG-500
Interacting with the admin menu to reach content creation and edit page stopped working during review. Manually tested if a regular user could reach the admin/content page, which was not possible. Therefore I refactored the code to go straight to the admin/content page instead of interacting with the menu to solve the issue.